feat: check for auth expiration and renew or logout
